Renamed PathContour to PathProfileContour for consistency.
